requirements.txt
pyspark
kafka-python
cassandra-driver
apache-airflow
streamlit
pandas
